Add ischanged.py.
This program takes input from stdin, hashes it, then compares it with
the last known value and if different, stores it and returns 1.
Otherwise it returns 0. This is intended to do things like check the
output of a process for changes. For instance you could curl a web site
regularly and trigger an alert if the output changes. Since it only
stores a hash (sha256) it's pretty lightweight.